requirements.txt
streamlit==1.29.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
Pillow==10.1.0